Move helper code to a helper module
